Fix price for doubles instead of int

+++ src/com/kennethswenson/gui/AddProductController.java
@@ -238,7 +238,7 @@ public class AddProductController {
                 totalPartsCost += part.getPrice();
             }
             try{
-                if(Integer.valueOf(tbPrice.getText()) < totalPartsCost){
+                if(Double.valueOf(tbPrice.getText()) < totalPartsCost){
                     valid = false;
                     sb.append("\nProduct cost must be more than the cost of the parts associated");
                     tbPrice.pseudoClassStateChanged(invalid, true);
